requirements.txt
pandas
scikit-learn
joblib
azureml-sdk
